+++ cpp/LeetCode/Problems/Array/Easy/LengthOfLastWord.hpp
@@ -14,7 +14,7 @@ namespace LengthOfLastWord {
 
     int lengthOfLastWord(string s) {
         int ans = 0;
-        for(int i = s.size() - 1; i >= 0; --i )
+        for(int i = (int)s.size() - 1; i >= 0; --i )
             if(s[i] != ' ') ++ans;
             else if(ans > 0) break;
